+++ inst/doc/example.R
@@ -1,0 +1,5 @@
+## ------------------------------------------------------------------------
+data("ChickWeight")
+ChickWeight <- dplyr::mutate(ChickWeight, Time = Time+1)
+ChickWeight <- NULL
+