+++ NEWS.md
@@ -1,4 +1,10 @@
 
+
+# seplyr 0.5.7 2018-06-30
+
+ * Remove use of isFALSE().
+ * Add gather_se() and spread_se().
+ 
 # seplyr 0.5.6 2018-06-23
 
  * Adjust dependencies.