+++ NEWS.md
@@ -1,4 +1,10 @@
 
+# seplyr 0.8.6 2020-04-30
+
+ * Patches for dplyr 1.0.0
+ * Improve tests
+ * Badges
+
 # seplyr 0.8.5 2020-01-16
 
  * Add grouped extend.